! The time_taken column for the scheduled task log is a float not an int. (ScheduledTasks.php, ManageMaintenance.php) [Bug 1321]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@6560 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ Sources/ManageMaintenance.php
@@ -341,7 +341,7 @@ function ScheduledTasks()
 				$total_time = round(array_sum(explode(' ', microtime())) - array_sum(explode(' ', $start_time)), 3);
 				$smcFunc['db_insert']('',
 					'{db_prefix}log_scheduled_tasks',
-					array('id_task' => 'int', 'time_run' => 'int', 'time_taken' => 'int'),
+					array('id_task' => 'int', 'time_run' => 'int', 'time_taken' => 'float'),
 					array($row['id_task'], time(), $total_time),
 					array('id_task')
 				);
